Move LanguageServerProcess to lib/server-manager.js
